Move ZipHandlerTest to correct package
